Added Character blueprints for the other npcs, Added a sleeping mechanic that can be built off under BP_bed
Added all characters in with 2 new ABPs so that we can use either charon's animations or Hades on nyx or the gorgon. they all have character blueprints because I didn't want to mess with other people's systems so this should hopefully allow them to add in the characters' animations without needing to affect the main ABP. The sleeping mechanic is on BP_Bed and it was made for the animations but can easily be added to if there isn't already a sleeping mechanic (I couldn't find one).